+++ tests/acl/test_mappers.py
@@ -20,8 +20,8 @@ Test suite for request-to-target mappers.
 
 from nose.tools import eq_, assert_raises, raises
 
-from repoze.what.acl.mappers.base import Mapper, CompoundMapper, Target, \
-                                         MappingError, NoTargetFoundError
+from repoze.what.exc import MappingError, NoTargetFoundError
+from repoze.what.acl.mappers.base import Mapper, CompoundMapper, Target
 from repoze.what.acl.mappers.pathinfo import PathInfoMapper
 from repoze.what.acl.mappers.routing_args import PositionalArgsMapper, \
                                                  NamedArgsMapper, \