+++ docs/source/News.rst
@@ -31,6 +31,7 @@ This document describes the releases of :mod:`repoze.what`.
     request and the new repoze.what "helpers" dict.
   * :meth:`repoze.what.predicates.Predicate.parse_variables` is gone because
     of the use of WebOb.
+  * Moved all the exceptions to :mod:`repoze.what.exc`.
 * Introduced :meth:`repoze.what.predicates.Predicate.is_met`.
 * Now :meth:`repoze.what.predicates.Predicate.unmet` receives an optional
   argument to override the error message.